+++ omgeo/services/__init__.py
@@ -1,5 +1,5 @@
 from .bing import Bing
-from .esri import EsriEUSoap, EsriEU, EsriNASoap, EsriNA, EsriWGS, EsriWGSSSL
+from .esri import EsriWGS
 from .us_census import USCensus
 from .mapquest import MapQuest, MapQuestSSL
 from .nominatim import Nominatim